+++ golded3/geread2.cpp
@@ -290,8 +290,8 @@ void GotoMsgno() {
     char buf[150];
     char* ptr = buf;
     int list_max = reader_msg->link.list_max();
-    ulong* replies = (ulong*)throw_calloc(list_max+1, sizeof(ulong));
-    ulong replyto, replynext;
+    uint32_t* replies = (uint32_t*)throw_calloc(list_max+1, sizeof(uint32_t));
+    uint32_t replyto, replynext;
     if(CFG->switches.get(disprealmsgno)) {
       ptr += sprintf(ptr, " %-5.5s:       %s %u", LNG->MsgReal, LNG->of, AA->Msgn.Count());
       replyto = reader_msg->link.to();
